Refactor: Update global dialog handling by replacing event-bus with subscribers and improve error handling in saveGlobalClasses function
