+++ packages/packages/core/editor-global-classes/src/components/class-manager/duplicate-label-dialog.tsx
@@ -27,7 +27,7 @@ export const DuplicateLabelDialog = ( {
 	onApprove?: () => void;
 } ) => {
 	const handleButtonClick = () => {
-		localStorage.setItem( 'elementor-global-classes-search', 'DUP_' );
+		localStorage.setItem( 'elementor-global-classes-search', DUP_PREFIX );
 		onApprove?.();
 		closeDialog();
 	};
