+++ packages/packages/libs/editor-ui/src/components/global-dialog/index.ts
@@ -1,2 +1,2 @@
 export { GlobalDialog } from './components/global-dialog';
-export { openDialog, closeDialog } from './event-bus';
+export { openDialog, closeDialog } from './subscribers';
